Move stats types into a separate module
